Add check to see if databases are online after clone

+++ bin/cleanup_clones.sh
@@ -12,20 +12,59 @@ do
 	echo "Removing service \"$CLONE\"."
 	docker service rm $CLONE
 done
-
-echo "Waiting for services to stop."
-sleep 10
+echo
 
 # We need to delete the volume on all nodes in the swarm.
-NODES=$( docker node ls | grep -v '*' | grep -v '^ID' | awk '{ print $2 }' )
+NODES=$( docker node ls | awk '/Active/ { print $(NF-3) }' | sort )
 
-echo "Removing volumes on this node."
-docker volume ls | grep percona_clone_ | awk '{ print $2 }' | xargs docker volume rm
+echo "Waiting for services to stop."
 for NODE in $NODES
 do
-	echo "Removing volumes on node \"$NODE\"."
-	ssh $NODE "docker volume ls | grep percona_clone_ | awk '{ print \$2 }' | xargs docker volume rm"
+	while true
+	do
+		ssh $NODE docker ps --filter status=running | grep -q percona_clone
+		if [[ ! $? = 0 ]]
+		then
+			echo "  - All services down on $NODE"
+			# Cleanup any remaining containers
+			COUNTER=0
+			EMPTY=0
+			while [[ $COUNTER -lt 10 ]]
+			do
+				ssh $NODE docker ps -a | grep -q percona_clone_
+				if [[ $? = 0 ]]
+				then
+					COUNTER=$(( $COUNTER + 1 ))
+				else
+					EMPTY=1
+					break
+				fi
+				sleep 1
+			done
+
+			if [[ $EMPTY = 0 ]]
+			then
+				echo "    - Removing old containers on $NODE"
+				ssh $NODE docker ps -a | grep percona_clone_ | awk '{ print $1 }' | xargs ssh $NODE docker rm
+			fi
+
+			break
+		fi
+
+		sleep 1
+	done
 done
 
+echo
+echo "Removing volumes."
+docker volume ls | grep -q percona_clone_ 
+if [[ $? = 0 ]]
+then
+	docker volume ls | grep percona_clone_ | awk '{ print $2 }' | xargs docker volume rm
+else
+	echo "  - No volumes found."
+fi
 
-# NETAPP
+
+
+
